🚧 chore(factories): add OptionFactory ✨ feat(factories): add bedrooms, floor, city, address, postal_code, and sold fields to PropertyFactory The OptionFactory has been added to the project. It defines the default state of the Option model, which includes a name field generated by the faker library.
The PropertyFactory has been updated to include new fields such as bedrooms, floor, city, address, postal_code, and sold. These fields are generated by the faker library and will be used to populate the Property model.

🌱 chore(DatabaseSeeder.php): seed options and properties
This commit adds the seeding of options and properties to the database seeder. 10 options are created using the Option model factory and 50 properties are created using the Property model factory. Each property is attached to 3 random options using the `hasAttached` method.